+++ tm-backend/.env
@@ -1,0 +1,5 @@
+SECRET_KEY="eac77e4e9a9a767b792779132e84ea37b1f4c31bec56714607f617a3fbdfbd53"
+ALGORITHM="HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES=30
+HOST="0.0.0.0"
+PORT=8008
