+++ tm-backend/app/database.py
@@ -1,0 +1,10 @@
+# Author Tom.Yang (https://github.com/7n8fail)
+
+users_db = {
+    "mockuser":{
+        "id": 1,
+        "username": "mockuser",
+        "email": "mock@user.com",
+        "password": "$2b$12$sErK932BEaLyIisz30PubepN7w91RLwkISWbAFYgUgoIqh8goJLEW",
+    }
+}
